DAG-main/OpenAI/requirements.txt
# OpenAI & LLM
openai==1.3.0
groq==0.4.1

# Data Processing
pandas==2.0.3
numpy==1.24.3
pydantic==2.4.2

# Web & Search
tavily-python==0.7.17
requests==2.32.5
beautifulsoup4==4.12.2
newspaper3k==0.2.8
feedparser==6.0.10
duckduckgo-search==3.9.10

# NLP & Search
sentence-transformers==2.2.2
rank-bm25==0.2.2
scikit-learn==1.3.0

# Finance
yfinance==0.2.28
polygon-io==1.9.1

# Async & Utilities
aiohttp==3.8.5
asyncio-contextmanager==1.0.0
python-dotenv==1.0.0
tenacity==8.2.3

# Logging & Monitoring
structlog==23.1.0
colorama==0.4.6

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1

# API
fastapi==0.104.1
uvicorn==0.24.0